Store the size and pointer of each allocation
In order for a conservative collector to trace the object graph it must
know the following two things about each potential root on the stack:
    1 The size of an allocated pointer's block
    2 How to get to the base pointer from a pointer inside an allocation
      block

This PR introduces a separate array which records this metadata on each
allocation. It uses the allocator in libc and provides an additional API
for a collector to query information about pointers at runtime.

In this PR, the allocation list is set arbitrarily at 2KB, and will
abort when full. A PR to fix this so that the list grows dynamically
will be forthcoming. Each metadata entry in the allocation list is 2
machine words and includes the allocated pointer, and the block size.

Insertion into the allocation list is done with an O(1) pointer bump
until it is fully initialised, at which point it becomes O(n) looking
for freed blocks as the list becomes fragmented. Removal of each entry
is O(n).