Make ALLOC_INFO thread-safe
System allocators are generally thread-safe, but gcmalloc reads and
writes metadata about each allocation to shared memory (ALLOC_INFO) upon
returning from the system allocator call. This requires additional
synchronisation mechanisms. It is not possible to use `sys::Sync::Mutex`
for this purpose as the implementation includes a heap allocation
containing a `pthread_mutex`. This would cause infinite recursion when
initialising the lock as the allocator would call itself.

Instead, we use a simple global spinlock built on an atomic bool to
guard access to the ALLOC_INFO list.