Update allocator to use the stabilised global_allocator API
Allocation in Rust is now done by calling the various methods on the
GlobalAlloc trait. gcmalloc now interacts with this API directly.

Annotating the struct which implements `GlobalAlloc` (in this case,
`ALLOCATOR`) with the `#[global_allocator]` attribute instructs the
runtime to perform all allocations through this trait implementation.
This means that pulling this crate into any Rust project will ensure
that allocation is performed via gcmalloc.