Record whether an allocation block is managed by gc
This is needed for a conservative collector which works in conjunction
with standard, non-gc'd memory alloc/dealloc. When it is time for the
collector to free unmarked memory, it must know which allocation blocks
are within its remit or else it may attempt to free memory which are not
intended to be GCd.