Add separate allocator for GC managed objects
This allocator records that a block is managed by the collector. A
separate allocation mechanism is needed for this purpose as it is not
possible to communicate the type information of an object to an
allocator. Thus to ensure that GC'd allocation block are known to the
GC, we must allocate all new GC objects with `GC_ALLOCATOR` instead of
the global allocator (`ALLOCATOR`).