Add a `Gc` smart pointer
This provides a convenience wrapper for the user to construct garbage
collected objects without having to fiddle with allocation details.
Eventually the interface will be similar to that of `Rc`.

For example, constructing a GC object using the constructor:

`let x = Gc::new(1234)`

The collector knows that it is responsible for managing the memory of
`x`.