Introduce explicit header for a `Gc<T>` container
The memory layout of a `Gc<T>` includes a single `usize` at a
`sizeof::<usize>` offset from the `objptr` value. In the future, this
will become necessary in order to store important properties about an
object such as its type and the mark-bit used by the collector. We make
this relationship explicit by giving it a proper type.