Hide GcHeader impl and use interior mutability
Using `Cell` for the header word in `Gc<T>` means that we don't have to
provide a mutable interface for it. Since the collector interacts with
Gc<T> objects unsafely with raw pointers in a few places, this approach
feels safer. Especially as rustc docs seem to suggest that data in
`Cell` is respected by the optimiser, and we don't have to worry so much
about strict aliasing in our own code.