Always run Valgrind and TSAN in gc_tests
It is now useful to explicitly test for the existence of memory
corruptions and invalid reads/writes. This is made much easier with
valgrind. We now run lang_tester in two phases: 1) with TSAN; and 2)
under Valgrind. Some tests now explicitly require valgrind to run. For
example, if we want to check that some memory was definitely freed, we
can try to read it later and assert that valgrind caught a
use-after-free. In cases like this, we add a filter to lang_tester to
exclude tests from phase 1 of testing if the file name contains the
substring "valgrind".