Dealloc `Gc` objects using their vtable info.
Storing the number of machine words to the baseptr in a `Gc` object's
header was unsafe. Not only did it omit alignment information, but, more
importantly, its effective integer range was less than `usize::MAX` as
we encoded the GC mark bit in the LSB of the header.

In this commit, we get around this problem by storing `Gc<T>` as a thin
trait object, where the header contains the pointer to the vtable. This
lets us take advantage of the fact that Rust stores the size and
alignment of a trait object as the second and third entires of a trait
object's vtable respectively.

During the sweep phase of collection, where we need to deallocate dead
objects, we can now reconstruct fat pointers to `Gc<T>` to get the size
and alignment of `T` without ever needing to know its concrete type
statically.

In order for this new storage scheme to work, `Gc<T>` needs to implement
some trait for all values of `T` so that it can be represented as a
statically known trait object. At present, the need for such a trait
does not exist, however, in the future a `Trace` trait will be necessary
for the GC to precisely trace through objects. We introduce this `Trace`
trait here as a placeholder with an empty `trace` method in the
meantime, so that we can take advantage of the vtable storage trick.