Add `VOH<T>` array type
A `VOH<T>` (vector of holes) is a contiguous growable array type
optimised for fast insertion and O(1) removal based on Rust's `RawVec`
low-level type.

`RawVec` already takes care of many of the low-level corner cases
involved in allocationg contiguous chunks on the heap.

Unlike the `Vec<T>` type available in the standard library,
`VOH<T>` does not shift remaining elements to the left after
removing an item. Instead, a `None` value is swapped in its place,
effectively leaving a hole inside the vector.

Accessing elements in a `VOH<T>`, however, can be slower than a
`Vec<T>`, as elements are no longer strictly contiguous; fragmentation
will occur over time as removed elements increase the distance between
those which remain.

`VOH<T>` does not use the global allocator, so it can be safely
used inside the collector with the knowledge that its contents are not
recorded in the block metadata. This means there is no danger of object
resurrection.