Add linked list header to all heap allocated blocks
During collection, we need a way to traverse each heap allocated block
found from the rootset for further pointers. Previously this was done
using a global vector which stored metadata about each allocation. This
was brittle and lead to synchronisation problems.

This commit adds a header to each block with pointers to next and
previous blocks, allowing them to traversed during collection. A doubly
linked list is used so that deallocation is O(1) instead of O(n).

We also fold the header for GC objects up into the block header.
Although we take a memory hit in doing this with an extra 4 machine
words necessary per each allocation, the implementation is simplified
and easier to debug in the meantime.