Rethink VOH with a simpler API and better name: GcVec
Since VOH is not needed for allocation anymore, its API can be
simplified so that it stores elements of type `T` instead of
`Option<T>`. For simplicity, we also remove the ability to delete
elements which are not at the end of the list.

VOH is still useful for GC internals because it is not tracked by the
global allocator, so we give it a more appropriate name: GcVec.