Add a preparation phase to reset mark bits
It turns out that with the benefit of hindsight, things are greatly
simplified if resetting mark bits is a distinct phase which takes place
before marking. Not only does this let us de-couple the collector from
much of the Gc and alloc APIs, but we also no longer need to lock the
collector on each fresh `Gc::new()` to see which state represents black.