Look for roots in static variables
Since statics can be mutable, they may contain roots to garbage
collected values. We identify the range where these roots can live in
the .data segment and scan it for pointers in a similar manner to the
stack. This is platform specific and we only provide a mechanism for
doing this in linux at the moment.