Add basic support for Gc<T: ?Sized>
In order to have the same flexibility as other smart pointer types, it
should be possible for users to store DSTs inside a Gc container. To do
this, we use the magic `CoerceUnsized` trait to convert a concrete type
to an unsized one when requested by the user at a "coercion point".

For example, lets imagine we have a trait `MyTrait` and a struct `S`,
where `S: MyTrait + Sized`. We might want to store a garbage collected
reference to a bare trait object as below:

    let gc: Gc<dyn MyTrait> = Gc::new(s) // where `s: S`.

This now compiles successfully because of our `CoerceUnsized`
implementation on `Gc`. It works because the compiler can see a valid
coercion from a concrete type to a DST: more specifically, the concrete
type `S` can be converted to the trait object `dyn MyTrait`, because `S`
is bound by the trait `MyTrait`.