Add `Any` type support to Gc<T>
The `Any` type provides a very restricted form of dynamic typing in
Rust. This means that a `Gc<T>` (where T: Any) can now be downcast to
its concrete type. `Any` also allows for testing whether values are of a
specified concrete type.

The `Any` trait does not, however, allow the user to see which traits
the `Gc<T>` implements.