Drop the marking worklist after mark phase
Previously, the marking worklist was a field on the `Collector` struct.
This meant that the blocks it held could be accessed through methods
used during the sweep and finalization phases. However, during these
latter stages, there are no guarantees that the blocks inside the
marking worklist are still live. This can lead to use-after-frees if one
is not careful to respect the implicit liveness invariants imposed by a
STW mark-sweep algorithm.

Since this is not ideal, we try to make some of these requirements
explicit. First, by constructing the worklist inside the mark phase and
passing it around by reference, we will now get compiler warnings if
someone tries to hold on to it in the Collector.

Second, the explicit lifetime annotations on a `Block` now means that we
can't accidentally pass `Block`s between different GC phases. More
specifically, a `Block` found when scanning the heap during the mark
phase will have a 'mrk lifetime. If we attempt to put this in the drop
queue for finalization, we will get a liftime error, as the drop queue
requires `Block<'static>`. Instead, we must re-obtain the `Block`
reference with the correct 'static lifetime by rescanning the heap
during the sweep phase.