Cache heap block pointer identification across a collection
During marking, pointer identification is slow because it needs to
traverse a linked-list spanning the entire heap while looking for a
pointer. We now sort and cache block pointers in-use by the mutator when
a collection is triggered so that they can be looked up in O(logn) using
binary search.

Unfortunately this required quite a bit of code shuffling to keep the
borrow checker happy from the previous couple of PRs.