Allow construction of uninitialized Gc values
This lets us create a Gc value to a `Gc<MaybeUninit<T>>` with a new constructor:

```rust pub fn new_from_layout<T>(layout: Layout) -> Option<Gc<T>> ```

The constructor returns `None` if `layout` is less than `size_of::<T>()` because
this could accidentally trigger out-of-bounds accesses.

Since a `Gc<MaybeUninit<T>>` will not be dropped until it is initialized, we
also need to modify the sweep and finalization phases to make sure of this.