Add mechanism for specifying how the GC should trace values
The collector currently treats all objects it encounters as opaque
blocks of possible pointer values. It then scans through these,
word-by-word, checking whether these possible pointers extend the object
graph. This approach is very simple, but it can miss pointers if they
are stored 'opaquely'. In other words, pointers which are stored in a
way which require some form of de-obfuscation before they can be
dereferenced will be missed by the collector.

To mitigate this, we need a way for the user to be able to tell the gc
that certain fields require de-obfuscating before they can be treated as
pointers. In this commit, we flesh out such an API.

The basic idea is to introduce a `Trace` trait for types to implement if they
contain opaque pointers. It defines as single `trace` method which
specifies how pointer fields can be de-obfuscated by the collector.

The `Trace` trait should not be implemented directly by the user,
however. Instead, we provide a custom derive attribute which we expect
users to use. Consider the following struct:

    struct S {
        opaque_ptr: usize,
        no_ptr: usize,
        normal_ptr: *mut Vec<...>
    }

If the collector traced `S` conservatively, it would miss this pointer
stored in the `opaque_ptr` field. To tell the collector how to
de-obfuscate it, we derive `Trace` and use a field level attribute as
shown below:

    #[derive(Trace)]
    struct S {
        #[trace = "path::to::deobfuscation::fn"]
        opaque_ptr: usize,
        no_ptr: usize,
        #[trace]
        normal_ptr: *mut Vec<...>
    }

The user needs to write a function to de-obfuscate each opaque field,
and then link it to the required field in the with the attribute:
`#[trace = "..."]`. The de-obfuscation function should take a single
argument, whose type is that of the field and return a pointer with the
type *const u8. For example, a de-obfuscation function for
`opaque_field` above might look like this:

    fn deobf_opaque_ptr(field: usize) -> *const u8 {
        (field | SOME_MASK) as *const u8
    }

Since a type which implements `Trace` will no longer be conservatively
traced, the user must also annotate other pointer-like fields which lead
to Gc values. This is done as above with the `#[trace]` attribute. This
is clearly error-prone and undesirable, and in the future this will not
be necessary.

This is insufficient to describe all types which may need a custom
tracing mechanism. However, I think it is quite ergonomic for the
common-case, and it can be built upon for more complex trace semantics.

The GC code required to respect this `Trace` api has not yet been
implemented.