Add lang_tester framework for Miri tests
Unfortunately this is a bit brittle because of limitations of the
cargo-miri binary [1]. To mitigate this, we create a temp directory and
rsync over the src files to compile a separate gcmalloc target with miri
flags. We then copy over the tests one at a time in order to run them
under Miri.

[1]: https://github.com/rust-lang/miri/issues/1173